requirements.txt
asgiref==3.10.0
attrs==25.4.0
autoflake==2.3.1
black==25.9.0
certifi==2025.10.5
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
distlib==0.4.0
dj-database-url==3.0.1
Django==5.2.7
django-extensions==4.1
django-oauth-toolkit==3.1.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
filelock==3.19.1
flake8==7.3.0
flake8-bugbear==24.12.12
flake8-comprehensions==3.17.0
git-filter-repo==2.47.0
identify==2.6.15
idna==3.11
inflection==0.5.1
iniconfig==2.1.0
isort==6.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jwcrypto==1.5.6
mccabe==0.7.0
mypy_extensions==1.1.0
nodeenv==1.9.1
oauthlib==3.3.1
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
pre_commit==4.3.0
psycopg==3.2.11
psycopg-binary==3.2.11
pycodestyle==2.14.0
pycparser==2.23
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
python-decouple==3.8
pytokens==0.1.10
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.28.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.34.0
